Began the process of adding commentary prose.
This process is going to take a while. I am giving full rein to my
verbose nature as I write my thoughts on the configuration file. I
have added a structure which I may use for the document as a
whole. The aim is to end up with a sensible structure where it should
be clear where configuration for a particular type will be found. This
may involve some moving stuff around.

I will also look to add more commentary (especially on document
structure) in the readme file.